% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{adding}
\alias{adding}
\alias{add_cur_head}
\alias{add_subhead}
\alias{add_custom}
\alias{add_pause}
\alias{add_br}
\alias{add_lb}
\alias{get_end}
\title{Additions to the document}
\usage{
add_cur_head(n = NULL, x = "")

add_subhead(n = NULL, sep = " - ", x = "", sub = TRUE)

add_custom(x, end = "pbr")

add_pause(br = TRUE, trailing = TRUE)

add_br(trailing = TRUE)

add_lb()

get_end(end, ...)
}
\arguments{
\item{n}{The heading level to use. Set to \code{NA} to ommit. Set to \code{NULL} (the
default) to use the value of \code{head_lv}.}

\item{x}{The custom text to add.}

\item{sep}{If \code{sub = TRUE}, the separator to use between the section headers.}

\item{sub}{Should the header be added to the current section header, or be a
standalone header?}

\item{end}{The type of end to add. See \code{get_end()}.}

\item{br}{Should a \code{'<br>'} be added after the pause?}

\item{trailing}{Should an empty line be added before?}

\item{...}{Arguments to pass to the \verb{add_*} functions.}
}
\description{
Functions to add content to de document:
\itemize{
\item \code{add_cur_head()} adds the heading of the current section, with the
specified level, and customized text. \code{add_subhead} is passed to a 'headx'
block, to add it to the doc,
with a new heading level, and possibly accompanying the section header..
\item \code{add_custom()} adds any custom text.
\item \code{add_pause()} adds \code{'. . .'}, \code{add_br()} adds \code{'<br>'}, and \code{add_lb()} adds
\code{'\\n\\n'}. \code{get_end()} is a wrapper to choose between them.
}
}
