% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{apply_splits}
\alias{apply_splits}
\title{Apply an Ordered Succession of Splits to Character Vector}
\usage{
apply_splits(doc_raw, split_args, sec_lv, clauses_pat)
}
\description{
Apply an Ordered Succession of Splits to Character Vector
}
\keyword{internal}
