% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{create_sections_names}
\alias{create_sections_names}
\title{Create pretty names for the sections and blocks}
\usage{
create_sections_names(sections)
}
\description{
Create pretty names for the sections and blocks
}
\keyword{internal}
