% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{editing}
\alias{editing}
\alias{e}
\alias{div}
\title{Edit Blocks}
\usage{
e(
  keep = NULL,
  end = NULL,
  breaks = NULL,
  adds = NULL,
  subs = NULL,
  modify = function(x) x,
  sep = add_pause(),
  sep_n = 1
)

div(
  keep = NULL,
  end = NULL,
  breaks = NULL,
  adds = NULL,
  subs = NULL,
  modify = function(x) x,
  sep = add_pause(),
  sep_n = NULL,
  sep_fragment = FALSE
)
}
\arguments{
\item{keep}{Indexes of the clauses (for \code{e()}) or lines (for \code{div()}) to keep.
Defaults to all (\code{NULL}).}

\item{end}{Character specifying what to add at the end of the block, passed
to \code{get_end()}. The default value is whatever was passed to the
\code{morph_sec()}'s \code{end} argument. Possible values are:
\itemize{
\item \code{'pbr'} for \code{add_pause()} (the default);
\item \code{'p'} for \code{add_pause(FALSE)};
\item \code{'br'} for \code{add_br()};
\item \code{'lb'} for \code{''} (an empty line);
\item \code{'none'} for \code{NULL} (nothing).
}}

\item{breaks}{Indexes to insert \code{sep} after. If \code{sep_fragment = TRUE}, it
should be a list of indexes to surround with '.fragment' markdown blocks.}

\item{adds, subs}{Character vector specifying additions or substitutions
(respectively) on the final punctuation of each clause in \code{keep}. Defaults
to none (\code{NULL}).}

\item{modify}{Function to to the block, after the base manipulations.}

\item{sep}{What to add after the indexes specified by \code{breaks}.}

\item{sep_n}{Add \code{sep} after the break + \code{sep_n[1]} elements. \code{sep_n[2]}
is for before the break (when \code{sep_fragment = TRUE}).}

\item{sep_fragment}{Instead of \code{sep}, should the elements specified by
\code{breaks} be encapsulated in a ".fragment" markdown block? (for revealjs).}
}
\description{
These functions are assigned to blocks inside \verb{plan_sec(b* = fun(...))}.
See \link[morphdown]{morphing}. The functions \code{e()} and \code{div()} edit
existing blocks:
\itemize{
\item \code{e()} is used on single-lines of text, that have been split into clauses.
\item \code{div()} is used on multi-line blocks of text, that have been split by
lines.
}
}
