% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{get_split_args}
\alias{get_split_args}
\title{Function to get the standard \code{split_args}}
\usage{
get_split_args(sec_lv, clauses_pat)
}
\description{
Function to get the standard \code{split_args}
}
\keyword{internal}
