% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_methods.R
\name{groups_from_indexes}
\alias{groups_from_indexes}
\title{Get a factor vector of 'group ids' from binary vector of 'regions'}
\usage{
groups_from_indexes(x, indexes, final, single)
}
\description{
Get a factor vector of 'group ids' from binary vector of 'regions'
}
\keyword{internal}
