% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_methods.R
\name{index_add_border}
\alias{index_add_border}
\title{Update the index of a match to \code{n} places before and/or after the match}
\usage{
index_add_border(x, border, type)
}
\description{
Update the index of a match to \code{n} places before and/or after the match
}
\keyword{internal}
