% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphdown-package.R
\docType{package}
\name{morphdown-package}
\alias{morphdown}
\alias{morphdown-package}
\title{morphdown: Programmatically Change R Markdown Files}
\description{
Programmatically change the structure of markdown files. Process a markdown file into a comprehensible list of blocks, and apply changes to each, through a plan that is efficient and robust to changes in the original file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/ricardo-semiao/morphdown}
  \item \url{https://ricardo-semiao.github.io/morphdown/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/morphdown/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
