% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{morphing}
\alias{morphing}
\alias{morph_doc}
\alias{morph_sec}
\title{Morph documents}
\usage{
morph_doc(sections, ..., head_lv = 2, end = "pbr")

morph_sec(..., head_lv = NULL, end = NULL)
}
\arguments{
\item{sections}{A list of sections as returned by \code{split_sections}.}

\item{...}{\verb{s* = morph_sec(...)} arguments for \code{morph_doc()}, or
\verb{b* = fun(...)} arguments for \code{morph_sec()}.}

\item{head_lv}{An integer value to set the heading level of the current
section header. The default value for \code{morph_sec} is whatever was passed to
the \code{morph_doc()}'s argument. Can be overridden by passing a \code{head1}
argument.}

\item{end}{To set the default value for \code{e()}'s' and \code{div()}'s \code{end}
argument. The default value for \code{morph_sec} is whatever was passed to
the \code{morph_doc()}'s argument.}
}
\value{
For \code{morph_doc()}, a character vector containing the morphed
document. \code{morph_sec()} should not be used outside \code{morph_doc()}.
}
\description{
These functions are used to programmatically alter the Rmd, once it has been
split (see \link[morphdown]{splitting}).

In \code{morph_doc()}, global changes can be defined. Then, the user must supply
one \verb{s* = morph_sec(...)} arguments, where \code{*} corresponds to the section
number, for each section he wants to keep.

In \code{morph_sec()}, local section changes can be defined. Then, the user must
supply one \verb{b* = fun(...)} to each block he wants to keep. There is also the
possibility to add new lines of markdown with unnamed calls. See the
available block editing functions in \link[morphdown]{editing}.
}
