% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{planning}
\alias{planning}
\alias{plan_doc}
\alias{plan_sec}
\title{Plan Document Morphing}
\usage{
plan_doc(sections, level_add = 0, ...)

plan_sec(head_level = NULL, ...)
}
\arguments{
\item{sections}{A list of sections as returned by \code{split_sections}.}

\item{level_add}{An integer value to adjust the header level of all sections.
For instance, a \code{level_add = 1} transform "# Header" into "## Header".
Negative values decrease the header value.}

\item{...}{\verb{s* = plan_sec(...)} arguments for \code{plan_doc}, or \verb{b* = fun(...)}
arguments for \code{plan_sec}.}

\item{head_level}{An integer value to set the heading level of the current
section header.}
}
\value{
For \code{plan_doc}, a character vector containing the processed document.
\code{plan_sec} should not be used outside \code{plan_doc}.
}
\description{
These functions are used to programmatically alter the Rmd, once it has been
split (see \link[morphdown]{splitting}).

In \code{plan_doc}, global changes can be defined. Then, the user must supply one
\verb{s* = plan_sec(...)} arguments, where \code{*} corresponds to the section
number, for each section he wants to keep.

In \code{plan_sec}, local section changes can be defined. Then, the user must
supply one \verb{b* = fun(...)} to each block he wants to keep. There is also the
possibility to add new lines of markdown with unnamed calls. See the
available block editing functions in \link[morphdown]{editing}.
}
