% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_methods.R
\name{split_between_12}
\alias{split_between_12}
\title{Split character vector between opening (1) and closing (2) matches}
\usage{
split_between_12(x, pat, border = c(0, 0), final)
}
\description{
Split character vector between opening (1) and closing (2) matches
}
\keyword{internal}
