% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_methods.R
\name{split_on_repeated}
\alias{split_on_repeated}
\title{Split character vector on groups of repeated matches}
\usage{
split_on_repeated(x, pat, border = c(0, 0), final)
}
\description{
Split character vector on groups of repeated matches
}
\keyword{internal}
