% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_methods.R
\name{split_on_single}
\alias{split_on_single}
\title{Split character vector on a single match}
\usage{
split_on_single(x, pat, final)
}
\description{
Split character vector on a single match
}
\keyword{internal}
