% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitting}
\alias{splitting}
\alias{split_sections}
\title{Split Rmd Into Sections and Blocks}
\usage{
split_sections(
  x,
  sec_lv = 3,
  replacements_pat = NULL,
  split_args = NULL,
  clauses_sep = ".:;!?"
)
}
\arguments{
\item{x}{A character vector containing the path to an Rmd file or the actual
text from an Rmd file.}

\item{sec_lv}{An integer specifying the maximum heading size (number)
of "#" to be considered. Sub-sections lower than that won't be split.}

\item{replacements_pat}{A named vector where names are patterns to match in
the document's text and values are the replacements for those patterns.
Passed to the \code{pattern} argument of \link[stringr]{str_replace_all}.
Defaults to no substitution.}

\item{split_args}{A list of 'splits' to separate the sections and blocks.
See the source code of \code{morphdown:::get_split_args()} for the default
splits.}

\item{clauses_sep}{A string with which characters to consider when splitting
the clauses of a text expression. Defaults to ".:;!?".}
}
\value{
A named list of character vectors. Each vector is a section named
\verb{s*}. Each section has blocks named \verb{b*}. Use those names to refer to each
in the plan functions.
}
\description{
This function takes a path to an R Markdown (.Rmd) file or a string vector of
text from an Rmd file, cleans it by replacing specified patterns, and then
splits it into sections and blocks. The result makes it easy to refer to each
section/block with the plan functions of the package.
}
