% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{subs_and_adds}
\alias{subs_and_adds}
\title{Additions and substitutions to each character element}
\usage{
subs_and_adds(kept, adds, subs)
}
\description{
Additions and substitutions to each character element
}
\keyword{internal}
