% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{editing}
\alias{editing}
\alias{e}
\alias{div}
\alias{add_cur_head}
\title{Edit Blocks}
\usage{
e(cs = NULL, subs = NULL, pat = NULL)

div()

add_cur_head(n = NULL)
}
\arguments{
\item{cs}{Integer vector of clauses to keep. Defaults to all.}

\item{subs}{Character vector specifying substitutions for the final
punctuation on each clause in \code{cs}. Defaults to none.}

\item{pat}{The pattern to break expressions into clauses. Defaults to
\code{"(?<=[\\\\.:;!?])"}.}

\item{n}{The heading level to use.}
}
\description{
These functions are assigned to blocks inside \verb{plan_sec(}b* = fun(...)\verb{)}.
See \link[morphdown]{planning}. \code{e} and \code{div} edit existing blocks, the rest
adds new markdown content.
\itemize{
\item \code{e} is used on single element blocks ('text expressions'). It splits the
expression into clauses, and allows the user to select and alter each.
\item \code{div} is used on multi-element blocks. Currently, it returns the block
as is.
\item \code{add_cur_head} adds the heading of the current section, with the specified
level.
}
}
